option_chain_system/requirements.txt
fyers-apiv3
pandas
python-dotenv
requests
pytz
apscheduler
psycopg2-binary
